Rename lib into src
