+++ src/command-instance.ts
